Implementations added

+++ src/HRMS.Domain/Entities/TimeSheets/TimeSheet.cs
@@ -11,9 +11,9 @@ namespace HRMS.Domain.Entities.TimeSheets
     public class TimeSheet: BaseAuditableEntity
     {
         public DateTimeOffset WorkingDay { get; set; }
-        public float WorkedHours { get; set; }
-
+        public double WorkedHours { get; set; }
         public Guid EmployeeId { get; set; }
         public Employee Employee { get; set; }
     }
+
 }
